debugged savedata, improved Tset
when you tried to save your data, you would get an error, I fixed this by using ; instead of : to express time.

I also changed Tset to accept an integer in regular notation instead of base exponential notation. before, the base exponent notation was set up so that 1e6 would give you a measurement time of .1 seconds, which was confusing.