debugged lag
lag was coming from the computer trying to catch up to the sr400 photon counter when the counter takes measurements quickely. when this happens, the computer spits out a multiple measurements at once, with all measurements but the last one being old data that has already been read. I created an if loop that ensures that only new data is being read.

I also added an option to either scroll and or autoscale the graph.
 scrolling the window clears old data off the graph. so it does not
experience lag. (lag mainly affects the time betwean stopping measuring and starting new measurement.) Scaling scales the window to fit all the data. it still becomes a little bit laggy when dealing with large quantities of data, but it is good to see thigs slow phenomena such as photo-bleaching in real time.